requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
views~=0.3
pip~=24.3.1
attrs~=22.2.0
wheel~=0.45.1
pillow~=10.4.0
filelock~=3.13.1
redis~=5.1.0
psycopg2-binary>=2.9.9
dj-database-url>=2.0.0